fix(cli): clear the four clippy warnings, two of them by writing the check

+++ crates/ackinacki-bridge/src/idempotency.rs
@@ -803,7 +803,7 @@ fn format_utc(secs: i64) -> String {
     let y = yoe as i64 + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
     let mp = (5 * doy + 2) / 153;
-    let d = (doy - (153 * mp + 2) / 5 + 1) as u32;
+    let d = doy - (153 * mp + 2) / 5 + 1;
     let m = if mp < 10 { mp + 3 } else { mp - 9 };
     let y = if m <= 2 { y + 1 } else { y };
